# api/openapi.yaml
openapi: 3.0.3
info:
  title: PGDock API
  version: 1.0.0
  description: |
    Control-plane API for PGDock. This file is the single source of truth:
    Go server interfaces (oapi-codegen) and TypeScript client types
    (openapi-typescript) are generated from it with `make generate`.
servers:
  - url: /
paths:
...
  /api/v1/settings/general:
    get:
      operationId: getGeneralSettings
      summary: DB hostname and pooler TLS status
      tags: [settings]
      responses:
        "200":
          description: Settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/db-host:
    put:
      operationId: putDbHost
      summary: Set the hostname clients use for databases
      description: Also starts obtaining the pooler's TLS certificate for it.
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbHostRequest"
      responses:
        "200":
          description: Updated settings.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneralSettings"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/settings/db-host/check:
    post:
      operationId: checkDbHost
      summary: Check that a hostname resolves to this server
      tags: [settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DbHostRequest"
      responses:
        "200":
          description: DNS check result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DnsCheck"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/operations/{id}:
    get:
      operationId: getOperation
      summary: Get one operation, including its step log
      tags: [operations]
      parameters:
        - $ref: "#/components/parameters/OperationID"
      responses:
        "200":
          description: The operation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Operation"
        default:
          $ref: "#/components/responses/Error"
...
  /api/v1/projects/{id}/settings:
    patch:
      operationId: updateProject
      summary: Update a project's name, description, or guardrails
      description: |
        Changes to guardrails that live on the backend or the pooler
        (connection limit, timeouts, pool size) queue an `apply_settings`
        operation, returned alongside the project.
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: The updated project.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUpdated"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    ProjectID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    OperationID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: not_implemented
        message:
          type: string
        quota:
          $ref: "#/components/schemas/QuotaItem"
    OperationStatus:
      type: string
      enum: [queued, running, succeeded, failed]
    OperationLogEntry:
      type: object
      required: [ts, step, level, msg]
      properties:
        ts:
          type: string
          format: date-time
        step:
          type: string
        level:
          type: string
          enum: [info, warn, error]
        msg:
          type: string
    Operation:
      type: object
      required: [id, kind, params, status, attempts, run_after, log, created_at]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          example: create
        project_id:
          type: string
          format: uuid
          nullable: true
        params:
          type: object
          additionalProperties: true
        status:
          $ref: "#/components/schemas/OperationStatus"
        attempts:
          type: integer
        run_after:
          type: string
          format: date-time
        log:
          type: array
          items:
            $ref: "#/components/schemas/OperationLogEntry"
        error:
          type: string
          nullable: true
        created_by:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
          nullable: true
...
    ProjectStatus:
      type: string
      enum: [provisioning, active, promoting, restoring, deleting, error]
    ProjectTier:
      type: string
      enum: [shared, dedicated]
    ConnectionInfo:
      type: object
      required: [host, session_port, pooled_port, database, user, sslmode, pooled_url, session_url]
      properties:
        host:
          type: string
          example: db.example.com
        session_port:
          type: integer
          example: 5432
        pooled_port:
          type: integer
          example: 6543
        database:
          type: string
          example: blog_k2f9
        user:
          type: string
          example: blog_k2f9_owner
        sslmode:
          type: string
          example: require
        pooled_url:
          type: string
          description: Transaction-mode URL (app traffic). Without a password unless returned at creation or rotation.
          example: postgresql://blog_k2f9_owner@db.example.com:6543/blog_k2f9?sslmode=require
        session_url:
          type: string
          description: Session-mode URL (migrations). Without a password unless returned at creation or rotation.
          example: postgresql://blog_k2f9_owner@db.example.com:5432/blog_k2f9?sslmode=require
    Project:
      type: object
      required: [id, org_id, name, slug, db_name, owner_role, tier, status, settings, created_at, connection]
      properties:
        id:
          type: string
          format: uuid
        org_id:
          type: string
          format: uuid
        my_role:
          $ref: "#/components/schemas/ProjectRole"
        storage_state:
          $ref: "#/components/schemas/StorageState"
        can_switch_credentials:
          type: boolean
          description: A V1 project that still uses its V1 owner role (V2 §10.2).
        legacy_credentials_until:
          type: string
          format: date-time
          nullable: true
          description: The V1 credentials stop working at this time.
        name:
          type: string
        slug:
          type: string
        db_name:
          type: string
        owner_role:
          type: string
        tier:
          $ref: "#/components/schemas/ProjectTier"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        description:
          type: string
          nullable: true
        settings:
          $ref: "#/components/schemas/ProjectSettings"
        created_at:
          type: string
          format: date-time
        connection:
          $ref: "#/components/schemas/ConnectionInfo"
        last_backup_at:
          type: string
          format: date-time
          nullable: true
          description: When the latest backup of this project finished.
        instance:
          $ref: "#/components/schemas/InstanceSummary"
        pitr_window:
          $ref: "#/components/schemas/PitrWindow"
        retired_copy_until:
          type: string
          format: date-time
          nullable: true
          description: After a promotion, when the read-only shared copy is dropped.
    ProjectSettings:
      type: object
      required: [connection_limit, pool_size, statement_timeout, idle_in_transaction_session_timeout, disk_warn_bytes, console_read_only]
      properties:
        connection_limit:
          type: integer
        pool_size:
          type: integer
        statement_timeout:
          type: string
        idle_in_transaction_session_timeout:
          type: string
        disk_warn_bytes:
          type: integer
          format: int64
        console_read_only:
          type: boolean
...
    DbHostRequest:
      type: object
      required: [db_host]
      properties:
        db_host:
          type: string
          example: db.example.com
    DnsCheck:
      type: object
      required: [host, addresses, server_addresses, points_here]
      properties:
        host:
          type: string
        addresses:
          type: array
          items:
            type: string
        server_addresses:
          type: array
          items:
            type: string
          description: Addresses this server considers its own.
        points_here:
          type: boolean
        error:
          type: string
          nullable: true
    TlsStatus:
      type: object
      required: [mode, state]
      properties:
        mode:
          type: string
          enum: [acme, self-signed, files, off]
        state:
          type: string
          enum: [ok, pending, error, off]
        host:
          type: string
          nullable: true
        issuer:
          type: string
          nullable: true
        not_after:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
    GeneralSettings:
      type: object
      required: [db_host, session_port, pooled_port, sslmode, tls]
      properties:
        db_host:
          type: string
        session_port:
          type: integer
        pooled_port:
          type: integer
        sslmode:
          type: string
        tls:
          $ref: "#/components/schemas/TlsStatus"
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          maxLength: 1000
          nullable: true
        settings:
          $ref: "#/components/schemas/ProjectSettingsPatch"
    ProjectSettingsPatch:
      type: object
      properties:
        connection_limit:
          type: integer
          minimum: 1
          maximum: 1000
        pool_size:
          type: integer
          minimum: 1
          maximum: 1000
        statement_timeout:
          type: string
          description: Postgres duration such as `60s`; empty to unset.
        idle_in_transaction_session_timeout:
          type: string
        disk_warn_bytes:
          type: integer
          format: int64
          minimum: 0
        console_read_only:
          type: boolean
    ProjectUpdated:
      type: object
      required: [project]
      properties:
        project:
          $ref: "#/components/schemas/Project"
        operation:
          $ref: "#/components/schemas/Operation"
...
    InstanceSummary:
      type: object
      required: [id, kind, status, node_id, node_name]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [shared, dedicated]
        status:
          type: string
        error:
          type: string
          nullable: true
        node_id:
          type: string
          format: uuid
        node_name:
          type: string
        profile:
          type: string
          nullable: true
        cpus:
          type: number
          nullable: true
        memory_mb:
          type: integer
          nullable: true
        volume_gb:
          type: integer
          nullable: true
    PitrWindow:
      type: object
      description: Dedicated only. Any time in [from, to] can be restored.
      required: [from, to]
      properties:
        from:
          type: string
          format: date-time
        to:
          type: string
          format: date-time
...
    ProjectRole:
      type: string
      enum: [admin, developer, read_only]
...
    QuotaItem:
      type: object
      required: [limit, used]
      properties:
        limit:
          type: string
          description: The limit's key (V2 §10.3), e.g. projects or project_storage_mb.
        used:
          type: number
        max:
          type: integer
          format: int64
          nullable: true
          description: Absent or null means unlimited.
...
    StorageState:
      type: string
      description: V2 §10.4 storage enforcement.
      enum: [none, warn, soft, hard]
...