# internal/store/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package store

import (
	"encoding/json"
	"net/netip"
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Alert struct {
	ID                 uuid.UUID
	Kind               string
	Key                string
	Severity           string
	TargetType         string
	TargetID           string
	TargetName         string
	Summary            string
	Detail             json.RawMessage
	Status             string
	StartedAt          time.Time
	LastSeenAt         time.Time
	ResolvedAt         *time.Time
	NotifiedAt         *time.Time
	ResolvedNotifiedAt *time.Time
	DeliveryAttempts   int32
	DeliveryError      *string
	DeliveryLock       *time.Time
}

type AuditLog struct {
	ID         int64
	UserID     *uuid.UUID
	Action     string
	TargetType *string
	TargetID   *string
	Detail     json.RawMessage
	Ip         *netip.Addr
	UserAgent  *string
	Outcome    string
	CreatedAt  time.Time
	ActorKind  string
	TokenID    *uuid.UUID
	OrgID      *uuid.UUID
	ProjectID  *uuid.UUID
	BreakGlass bool
}

type AuthChallenge struct {
	ID        string
	Kind      string
	UserID    *uuid.UUID
	Payload   []byte
	Attempts  int32
	ExpiresAt time.Time
	CreatedAt time.Time
}

type Backup struct {
	ID              uuid.UUID
	ProjectID       *uuid.UUID
	Kind            string
	ObjectKey       string
	SizeBytes       *int64
	Checksum        *string
	StartedAt       time.Time
	FinishedAt      *time.Time
	Status          string
	ExpiresAt       *time.Time
	StorageTargetID *uuid.UUID
	OperationID     *uuid.UUID
	KeyWrapped      []byte
	Error           *string
	DeletedAt       *time.Time
}

type BreakGlassSession struct {
	ID        uuid.UUID
	OrgID     uuid.UUID
	AdminID   uuid.UUID
	Reason    string
	StartsAt  time.Time
	ExpiresAt time.Time
	EndedAt   *time.Time
	EndedBy   *uuid.UUID
}

type DedicatedRequest struct {
	ID           uuid.UUID
	OrgID        uuid.UUID
	ProjectID    uuid.UUID
	RequestedBy  uuid.UUID
	Profile      json.RawMessage
	Reason       *string
	Status       string
	DecidedBy    *uuid.UUID
	DecidedAt    *time.Time
	DecisionNote *string
	CreatedAt    time.Time
}

type EmailToken struct {
	TokenHash string
	UserID    uuid.UUID
	Purpose   string
	ExpiresAt time.Time
	UsedAt    *time.Time
	CreatedAt time.Time
}

type Instance struct {
	ID          uuid.UUID
	NodeID      uuid.UUID
	Kind        string
	PgVersion   int32
	Port        int32
	ContainerID *string
	CpuLimit    pgtype.Numeric
	MemLimitMb  *int32
	VolumeGb    *int32
	Status      string
	CreatedAt   time.Time
	AdminHost   *string
	AdminPort   *int32
	Host        *string
	AdminSecret []byte
	Profile     *string
	WalgPrefix  *string
	Error       *string
	DeletedAt   *time.Time
	OrgID       *uuid.UUID
}

type Invitation struct {
	ID           uuid.UUID
	Email        string
	TokenHash    string
	Kind         string
	OrgID        *uuid.UUID
	OrgRole      *string
	ProjectRoles json.RawMessage
	InvitedBy    uuid.UUID
	ExpiresAt    time.Time
	AcceptedAt   *time.Time
	AcceptedBy   *uuid.UUID
	RevokedAt    *time.Time
	CreatedAt    time.Time
}

type MetricPoint struct {
	Scope      string
	ScopeID    uuid.UUID
	Metric     string
	Ts         time.Time
	Resolution string
	Value      float64
}

type Node struct {
	ID                    uuid.UUID
	Name                  string
	PrivateAddr           string
	AgentPort             int32
	Role                  string
	AgentCertFp           *string
	PgAdminSecret         []byte
	Capacity              json.RawMessage
	Status                string
	LastHeartbeat         *time.Time
	CreatedAt             time.Time
	AgentHost             *string
	AgentVersion          *string
	RegistrationToken     *string
	RegistrationExpiresAt *time.Time
	LastReachableAt       *time.Time
}

type Operation struct {
	ID         uuid.UUID
	Kind       string
	ProjectID  *uuid.UUID
	Params     json.RawMessage
	Status     string
	Attempts   int32
	RunAfter   time.Time
	LockedBy   *string
	LockedAt   *time.Time
	Log        json.RawMessage
	Error      *string
	CreatedBy  *uuid.UUID
	CreatedAt  time.Time
	FinishedAt *time.Time
}

type OrgMember struct {
	OrgID     uuid.UUID
	UserID    uuid.UUID
	Role      string
	CreatedAt time.Time
}

type Organization struct {
	ID                 uuid.UUID
	Name               string
	Slug               string
	PersonalOwnerID    *uuid.UUID
	PlanID             uuid.UUID
	LimitOverrides     json.RawMessage
	DedicatedAllowance json.RawMessage
	Settings           json.RawMessage
	Status             string
	SuspendedReason    *string
	OutboundDisabled   bool
	DeleteAfter        *time.Time
	CreatedAt          time.Time
	SuspendedAt        *time.Time
	DeleteRequestedBy  *uuid.UUID
}

type Project struct {
	ID                  uuid.UUID
	Name                string
	Slug                string
	DbName              string
	OwnerRole           string
	ScramVerifier       string
	Tier                string
	InstanceID          uuid.UUID
	Status              string
	Settings            json.RawMessage
	StorageTargetID     *uuid.UUID
	Extensions          []string
	Description         *string
	CreatedBy           *uuid.UUID
	CreatedAt           time.Time
	DeletedAt           *time.Time
	OrgID               uuid.UUID
	AliasDbName         *string
	LegacyOwnerRole     *string
	LegacyScramVerifier *string
	LegacyUntil         *time.Time
	StorageState        string
	StorageStateAt      *time.Time
}

type ProjectDbUser struct {
	ProjectID     uuid.UUID
	UserID        uuid.UUID
	OrgID         uuid.UUID
	RoleName      string
	ScramVerifier string
	Access        string
	CreatedAt     time.Time
	RotatedAt     *time.Time
}

type ProjectMember struct {
	ProjectID uuid.UUID
	UserID    uuid.UUID
	OrgID     uuid.UUID
	Role      string
	AddedBy   *uuid.UUID
	CreatedAt time.Time
}

type QuotaPlan struct {
	ID        uuid.UUID
	Name      string
	Limits    json.RawMessage
	CreatedAt time.Time
}

type ReapedSession struct {
	ID        int64
	ProjectID uuid.UUID
	OrgID     uuid.UUID
	Kind      string
	RoleName  string
	DurationS int32
	Query     *string
	CreatedAt time.Time
}

type RetiredDatabase struct {
	ID         uuid.UUID
	ProjectID  uuid.UUID
	InstanceID uuid.UUID
	DbName     string
	OwnerRole  string
	Reason     string
	DropAfter  time.Time
	DroppedAt  *time.Time
	CreatedAt  time.Time
}

type Session struct {
	ID         string
	UserID     uuid.UUID
	CreatedAt  time.Time
	LastSeenAt time.Time
	ReauthAt   *time.Time
	Ip         *netip.Addr
	UserAgent  *string
}

type Setting struct {
	Key       string
	Value     json.RawMessage
	UpdatedAt time.Time
}

type StorageTarget struct {
	ID          uuid.UUID
	Name        string
	Endpoint    string
	Bucket      string
	Prefix      string
	Credentials []byte
	IsDefault   bool
	CreatedAt   time.Time
}

type TermsAcceptance struct {
	UserID     uuid.UUID
	Version    int32
	AcceptedAt time.Time
	Ip         *netip.Addr
}

type TermsVersion struct {
	Version     int32
	TermsMd     string
	PrivacyMd   string
	PublishedBy *uuid.UUID
	PublishedAt time.Time
}

type UsageRecord struct {
	OrgID       uuid.UUID
	ProjectID   uuid.UUID
	Metric      string
	Granularity string
	PeriodStart time.Time
	Quantity    pgtype.Numeric
	PlanID      uuid.UUID
}

type User struct {
	ID              uuid.UUID
	Email           string
	PasswordHash    string
	TotpSecret      []byte
	PlatformRole    string
	CreatedAt       time.Time
	DisabledAt      *time.Time
	FailedLogins    int32
	LockedUntil     *time.Time
	TotpLastStep    int64
	Name            *string
	EmailVerifiedAt *time.Time
	ApprovedAt      *time.Time
	RecoveryCodes   []byte
	LastActiveAt    *time.Time
}